armax_backend-main/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.8.2
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-import-export==3.2.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.4.0
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
djangorestframework==3.14.0
